Exclude commit messages from vimStartup positioning
The vimStartup autocmd group restores the cursor position when opening a
file.  This is counter-productive for commit messages which are usually
recreated as blank commit messages before opening and the desired cursor
position is the default topmost position.

This has been fixed upstream.  Include the fix here to solve it for old
versions as well.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>