Replace Pathogen with Vim 8 packages
I've decided to bite the bullet and switch from Pathogen to Vim 8
packages for managing Vim plugins.  I've created a vim-7 branch for
cases where I need plugins in Vim 7, but I plan to use more plugins
which require Vim 8 features to be useful and don't see a need to keep
Vim 7 compatibility in the master branch.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>