Use syntastic/master with W3 support for SVG+XHTML
Now that support has been merged, there is no need to use my modified
barnch.  Use the master branch.

The default doctype changed for HTML and XHTML.  Configure them to empty
for HTML and XHTML so that they are auto-detected by the validator
(based on the DOCTYPE, Content-Type, and content).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>