Use vim-jsonc for JSON with comments
Since it appears unlikely that support for comments will be added to
vim-json (and, even if it is, using a variable to control support is
problematic for modelines and detection+configuration) I have created a
package for a filetype plugin for a new "jsonc" filetype.  Use it
instead of the patched version of vim-json.  Revert vim-json to upstream
master.

Also remove the ALE configuration since the plugin package contains a
jsonlint ALE linter (which doesn't require my custom cjsonlint script).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>