Improve NeoVim Wayland clipboard handling
Only set g:clipboard when $WAYLAND_DISPLAY is not empty, since the
default provider is preferable when not running under Wayland.

Improve error handling by removing the pipe to tr so we can check the
exit code of wl-paste.  When wl-paste exits non-zero, print rather than
paste the error message.

Split on \n or \r\n, so that \r characters are ignored when part of line
breaks, but not if embedded in text.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>